separate validating methods into a different file (just to cut down on file length...)
